Fix PR #268 ( → ) perf-gate failure: bench subprocess workers fail with  because PYTHONPATH does not include the ark build directory. Fix  in  to resolve the built ark package path (e.g. from the build tree or installed site-packages), then verify the bench produces real latency numbers on multi-GPU. Target PERF_GATE ratio ≤ 1.0×.
