+++ ark/codegen.cpp
@@ -314,14 +314,15 @@ std::string CodeGenerator::Impl::def_task(const Json &task_json) {
                 auto moff = arg.value<ModelOffset>();
                 size_t buffer_id = moff.buffer_id();
                 auto buf_info = buf_reg.get(buffer_id);
-                if ((buf_info && buf_info->is_external) ||
-                    (buffer_id_to_offset_.find(buffer_id) ==
-                     buffer_id_to_offset_.end())) {
-                    // External buffer or buffer not in the local
-                    // allocation map: offset relative to its own base.
+                if (buf_info && buf_info->is_external) {
+                    // External buffer: offset relative to its own base.
                     ss_desc << moff.value();
                 } else {
-                    size_t buffer_offset = buffer_id_to_offset_.at(buffer_id);
+                    auto it = buffer_id_to_offset_.find(buffer_id);
+                    if (it == buffer_id_to_offset_.end()) {
+                        ERR(InternalError, "buffer ID not found: ", buffer_id);
+                    }
+                    size_t buffer_offset = it->second;
                     size_t offset = buffer_offset + moff.value();
                     ss_desc << offset;
                 }
